+++ examples/simple_tcp_client.py
@@ -3,6 +3,7 @@
 This program sends 10 random values between 0.0 and 1.0 to the /filter address,
 and listens for incoming messages for 1 second between each value.
 """
+
 import argparse
 import random
 
@@ -10,12 +11,15 @@ from pythonosc import tcp_client
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("--ip", default="127.0.0.1",
-                        help="The ip of the OSC server")
-    parser.add_argument("--port", type=int, default=5005,
-                        help="The port the OSC server is listening on")
-    parser.add_argument("--mode", default="1.1",
-                        help="The OSC protocol version of the server (default is 1.1)")
+    parser.add_argument("--ip", default="127.0.0.1", help="The ip of the OSC server")
+    parser.add_argument(
+        "--port", type=int, default=5005, help="The port the OSC server is listening on"
+    )
+    parser.add_argument(
+        "--mode",
+        default="1.1",
+        help="The OSC protocol version of the server (default is 1.1)",
+    )
     args = parser.parse_args()
 
     with tcp_client.SimpleTCPClient(args.ip, args.port, mode=args.mode) as client: