+++ pythonosc/dispatcher.py
@@ -2,6 +2,7 @@
 """
 
 import collections
+import inspect
 import logging
 import re
 import time
@@ -11,6 +12,7 @@ from typing import (
     List,
     Union,
     Any,
+    AnyStr,
     Generator,
     Tuple,
     Callable,
@@ -18,6 +20,7 @@ from typing import (
     DefaultDict,
 )
 from pythonosc.osc_message import OscMessage
+from pythonosc.osc_message_builder import ArgValue
 
 
 class Handler(object):
@@ -53,23 +56,30 @@ class Handler(object):
             and self.needs_reply_address == other.needs_reply_address
         )
 
-    def invoke(self, client_address: Tuple[str, int], message: OscMessage) -> None:
+    def invoke(
+        self, client_address: Tuple[str, int], message: OscMessage
+    ) -> Union[None, AnyStr, Tuple[AnyStr, ArgValue]]:
         """Invokes the associated callback function
 
         Args:
             client_address: Address match that causes the invocation
             message: Message causing invocation
+        Returns:
+            The result of the handler function can be None, a string OSC address, or a tuple of the OSC address
+            and arguments.
         """
         if self.needs_reply_address:
             if self.args:
-                self.callback(client_address, message.address, self.args, *message)
+                return self.callback(
+                    client_address, message.address, self.args, *message
+                )
             else:
-                self.callback(client_address, message.address, *message)
+                return self.callback(client_address, message.address, *message)
         else:
             if self.args:
-                self.callback(message.address, self.args, *message)
+                return self.callback(message.address, self.args, *message)
             else:
-                self.callback(message.address, *message)
+                return self.callback(message.address, *message)
 
 
 class Dispatcher(object):
@@ -93,11 +103,17 @@ class Dispatcher(object):
 
         The callback function must have one of the following signatures:
 
-        ``def some_cb(address: str, *osc_args: List[Any]) -> None:``
-        ``def some_cb(address: str, fixed_args: List[Any], *osc_args: List[Any]) -> None:``
+        ``def some_cb(address: str, *osc_args: List[Any]) -> Union[None, AnyStr, Tuple(str, ArgValue)]:``
+        ``def some_cb(address: str, fixed_args: List[Any], *osc_args: List[Any]) -> Union[None, AnyStr,
+                                                                                          Tuple(str, ArgValue)]:``
 
-        ``def some_cb(client_address: Tuple[str, int], address: str, *osc_args: List[Any]) -> None:``
-        ``def some_cb(client_address: Tuple[str, int], address: str, fixed_args: List[Any], *osc_args: List[Any]) -> None:``
+        ``def some_cb(client_address: Tuple[str, int], address: str, *osc_args: List[Any]) -> Union[None, AnyStr,
+                                                                                                    Tuple(str, ArgValue)]:``
+        ``def some_cb(client_address: Tuple[str, int], address: str, fixed_args: List[Any], *osc_args: List[Any]) -> Union[None, AnyStr, Tuple(str, ArgValue)]:``
+
+        The callback function can return None, or a string representing an OSC address to be returned to the client,
+        or a tuple that includes the address and ArgValue which will be converted to an OSC message and returned to
+        the client.
 
         Args:
             address: Address to be mapped
@@ -204,7 +220,7 @@ class Dispatcher(object):
 
     def call_handlers_for_packet(
         self, data: bytes, client_address: Tuple[str, int]
-    ) -> None:
+    ) -> List:
         """Invoke handlers for all messages in OSC packet
 
         The incoming OSC Packet is decoded and the handlers for each included message is found and invoked.
@@ -212,8 +228,9 @@ class Dispatcher(object):
         Args:
             data: Data of packet
             client_address: Address of client this packet originated from
+        Returns: A list of strings or tuples to be converted to OSC messages and returned to the client
         """
-
+        results = list()
         # Get OSC messages from all bundles or standalone message.
         try:
             packet = osc_packet.OscPacket(data)
@@ -226,9 +243,84 @@ class Dispatcher(object):
                 if timed_msg.time > now:
                     time.sleep(timed_msg.time - now)
                 for handler in handlers:
-                    handler.invoke(client_address, timed_msg.message)
+                    result = handler.invoke(client_address, timed_msg.message)
+                    if result is not None:
+                        results.append(result)
         except osc_packet.ParseError:
             pass
+        return results
+
+    async def async_call_handlers_for_packet(
+        self, data: bytes, client_address: Tuple[str, int]
+    ) -> List:
+        """
+        This function calls the handlers registered to the dispatcher for
+        every message it found in the packet.
+        The process/thread granularity is thus the OSC packet, not the handler.
+
+        If parameters were registered with the dispatcher, then the handlers are
+        called this way:
+          handler('/address that triggered the message',
+                  registered_param_list, osc_msg_arg1, osc_msg_arg2, ...)
+        if no parameters were registered, then it is just called like this:
+          handler('/address that triggered the message',
+                  osc_msg_arg1, osc_msg_arg2, osc_msg_param3, ...)
+        """
+
+        # Get OSC messages from all bundles or standalone message.
+        results = []
+        try:
+            packet = osc_packet.OscPacket(data)
+            for timed_msg in packet.messages:
+                now = time.time()
+                handlers = self.handlers_for_address(timed_msg.message.address)
+                if not handlers:
+                    continue
+                # If the message is to be handled later, then so be it.
+                if timed_msg.time > now:
+                    time.sleep(timed_msg.time - now)
+                for handler in handlers:
+                    if inspect.iscoroutinefunction(handler.callback):
+                        if handler.needs_reply_address:
+                            result = await handler.callback(
+                                client_address,
+                                timed_msg.message.address,
+                                handler.args,
+                                *timed_msg.message,
+                            )
+                        elif handler.args:
+                            result = await handler.callback(
+                                timed_msg.message.address,
+                                handler.args,
+                                *timed_msg.message,
+                            )
+                        else:
+                            result = await handler.callback(
+                                timed_msg.message.address, *timed_msg.message
+                            )
+                    else:
+                        if handler.needs_reply_address:
+                            result = handler.callback(
+                                client_address,
+                                timed_msg.message.address,
+                                handler.args,
+                                *timed_msg.message,
+                            )
+                        elif handler.args:
+                            result = handler.callback(
+                                timed_msg.message.address,
+                                handler.args,
+                                *timed_msg.message,
+                            )
+                        else:
+                            result = handler.callback(
+                                timed_msg.message.address, *timed_msg.message
+                            )
+                    if result:
+                        results.append(result)
+        except osc_packet.ParseError as e:
+            pass
+        return results
 
     def set_default_handler(
         self, handler: Callable, needs_reply_address: bool = False