+++ pythonosc/test/test_tcp_client.py
@@ -6,7 +6,7 @@ from pythonosc import osc_message_builder, slip, tcp_client
 
 
 class TestTcpClient(unittest.TestCase):
-    @mock.patch('socket.socket')
+    @mock.patch("socket.socket")
     def test_client(self, mock_socket_ctor):
         mock_socket = mock_socket_ctor.return_value
         mock_send = mock.Mock()
@@ -16,12 +16,12 @@ class TestTcpClient(unittest.TestCase):
 
         mock_socket.sendall = mock_send
         mock_socket.recv = mock_recv
-        msg = osc_message_builder.OscMessageBuilder('/').build()
-        with tcp_client.TCPClient('::1', 31337) as client:
+        msg = osc_message_builder.OscMessageBuilder("/").build()
+        with tcp_client.TCPClient("::1", 31337) as client:
             client.send(msg)
         mock_socket.sendall.assert_called_once_with(slip.encode(msg.dgram))
 
-    @mock.patch('socket.socket')
+    @mock.patch("socket.socket")
     def test_simple_client(self, mock_socket_ctor):
         mock_socket = mock_socket_ctor.return_value
         mock_send = mock.Mock()
@@ -31,13 +31,13 @@ class TestTcpClient(unittest.TestCase):
 
         mock_socket.sendall = mock_send
         mock_socket.recv = mock_recv
-        with tcp_client.SimpleTCPClient('::1', 31337) as client:
-            client.send_message('/', [])
+        with tcp_client.SimpleTCPClient("::1", 31337) as client:
+            client.send_message("/", [])
         mock_socket.sendall.assert_called_once()
 
 
 class TestAsyncTcpClient(unittest.IsolatedAsyncioTestCase):
-    @mock.patch('asyncio.open_connection')
+    @mock.patch("asyncio.open_connection")
     async def test_send(self, mock_socket_ctor):
         mock_reader = mock.Mock()
         mock_writer = mock.Mock()
@@ -46,22 +46,22 @@ class TestAsyncTcpClient(unittest.IsolatedAsyncioTestCase):
         mock_socket_ctor.return_value = (mock_reader, mock_writer)
         loop = asyncio.get_running_loop()
         loop.set_debug(False)
-        msg = osc_message_builder.OscMessageBuilder('/').build()
-        async with tcp_client.AsyncOSCTCPClient('::1', 31337) as client:
+        msg = osc_message_builder.OscMessageBuilder("/").build()
+        async with tcp_client.AsyncTCPClient("::1", 31337) as client:
             await client.send(msg)
 
         self.assertTrue(mock_writer.write.called)
         mock_writer.write.assert_called_once_with(slip.encode(msg.dgram))
 
-    @mock.patch('asyncio.open_connection')
+    @mock.patch("asyncio.open_connection")
     async def test_send_message_calls_send_with_msg(self, mock_socket_ctor):
         mock_reader = mock.Mock()
         mock_writer = mock.Mock()
         mock_writer.drain = mock.AsyncMock()
         mock_writer.wait_closed = mock.AsyncMock()
         mock_socket_ctor.return_value = (mock_reader, mock_writer)
-        async with tcp_client.AsyncSimpleTCPClient('::1', 31337) as client:
-            await client.send_message('/address', 1)
+        async with tcp_client.AsyncSimpleTCPClient("::1", 31337) as client:
+            await client.send_message("/address", 1)
         self.assertTrue(mock_writer.write.called)